pyproject.toml
[project]
name = "actovator-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "copilotkit==0.1.61",
    "langchain==0.3.21",
    "langgraph>=0.6.5",
    "langchain-openai>=0.3.32",
    "langfuse>=3.3.4",
    "pydantic-settings>=2.10.1",
    "sqlmodel>=0.0.24",
    "structlog>=25.4.0",
    "uvicorn[standard]>=0.35.0",
    "langgraph-checkpoint>=2.1.1",
    "langgraph-checkpoint-postgres>=2.0.23",
    "psycopg2>=2.9.10",
    "psycopg[binary]>=3.2.10",
    "fastapi[standard]>=0.115.14",
    "langchain-chroma>=0.2.6",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.182.0",
]
[tool.poetry]
packages = [{ include = "app" }]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
